// apps/backend/src/constant/operation-type.ts
export enum OperationCategory {
  AUTH = "AUTH",
  USER_MANAGEMENT = "USER_MANAGEMENT",
  PERMISSION = "PERMISSION",
  GROUP_MANAGEMENT = "GROUP_MANAGEMENT",
  SYSTEM = "SYSTEM",
  RELAY = "RELAY",
  BILLING = "BILLING",
  OJ_SUBMITTER = "OJ_SUBMITTER",
  SECURITY = "SECURITY",
  JSON_ENDPOINT = "JSON_ENDPOINT",
  ARTICLE = "ARTICLE",
  LEGAL_POLICY = "LEGAL_POLICY",
  NOTIFICATION = "NOTIFICATION",
}